requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
boto3==1.26.142
botocore==1.29.142
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-redis==5.2.0
django-storages==1.11.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
